py/requirements.txt
discord.py
flask
flask_cors
python-dotenv
pyglet